# project-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Provider Object Storage API
  version: 0.1.0
  description: |
    Initial contract for the implemented control-plane and custom storage API.
    This draft covers the organization membership and role administration
    vertical slice together with provider setup and storage clients.
    Organization management routes require a user JWT and the
    x-organization-id header. Platform routes require a global role with the
    platform permission and do not accept an organization header as scope.
    Storage routes require an API credential in the form keyId.secret and
    enforce namespace and folder-grant boundaries.
servers:
  - url: http://localhost:4000
    description: Local API
...
paths:
...
  /v1/providers/{providerId}/test:
    post:
      tags: [Providers]
      operationId: testProvider
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                bucketName:
                  type: string
                  minLength: 3
                  maxLength: 255
      responses:
        '200':
          description: Provider test succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderTestResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/ProviderFailure'
...
  /v1/api-credentials:
    get:
      tags: [API credentials]
      operationId: listApiCredentials
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Credential metadata; secrets are never returned by this operation.
          content:
            application/json:
              schema:
                type: object
                required: [credentials]
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiCredential'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [API credentials]
      operationId: createApiCredential
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiCredentialCreateRequest'
      responses:
        '201':
          description: Credential metadata and one-time secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedApiCredential'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/api-credentials/{credentialId}/rotate:
    post:
      tags: [API credentials]
      operationId: rotateApiCredential
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CredentialId'
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiCredentialRotateRequest'
      responses:
        '200':
          description: Rotated credential metadata and one-time secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedApiCredential'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/api-credentials/{credentialId}/revoke:
    post:
      tags: [API credentials]
      operationId: revokeApiCredential
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CredentialId'
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Revoked credential metadata.
          content:
            application/json:
              schema:
                type: object
                required: [credential]
                properties:
                  credential:
                    $ref: '#/components/schemas/ApiCredential'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived dashboard access token.
    storageCredential:
      type: http
      scheme: bearer
      bearerFormat: keyId.secret
      description: API credential scoped to one organization and namespace.
    metricsBearer:
      type: http
      scheme: bearer
      description: Deployment-provided metrics scrape token.
  parameters:
    OrganizationIdHeader:
      name: x-organization-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CredentialId:
      name: credentialId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Missing, invalid, expired, revoked, or out-of-scope credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource was not found in the caller's tenant scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ProviderFailure:
      description: Provider call failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
...
    ProviderTestResult:
      type: object
      required: [ok, bucketNames, capabilities]
      properties:
        ok: { type: boolean }
        bucketNames:
          type: array
          items: { type: string }
        capabilities:
          type: array
          items:
            type: object
            required: [capability, supported]
            properties:
              capability: { type: string }
              supported: { type: boolean }
              details:
                type: object
                additionalProperties: { type: string }
...
    ApiCredential:
      type: object
      required: [id, keyId, status, createdAt, updatedAt, createdById, expiresAt, revokedAt, lastUsedAt, namespace]
      properties:
        id: { type: string, format: uuid }
        keyId: { type: string }
        status: { type: string, enum: [ACTIVE, REVOKED, EXPIRED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdById: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        namespace:
          type: object
          required: [id, name, slug]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
    ApiCredentialCreateRequest:
      type: object
      required: [namespaceId]
      additionalProperties: false
      properties:
        namespaceId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time, nullable: true }
    ApiCredentialRotateRequest:
      type: object
      additionalProperties: false
      properties:
        expiresAt: { type: string, format: date-time, nullable: true }
    CreatedApiCredential:
      type: object
      required: [credential, secret]
      properties:
        credential: { $ref: '#/components/schemas/ApiCredential' }
        secret:
          type: string
          writeOnly: true
          description: Returned only at create/rotate time; store it immediately.
...